requirements.txt
colorama==0.4.3
comtypes==1.1.7
et-xmlfile==1.0.1
gast==0.3.3
grpcio==1.31.0
h5py==2.10.0
imutils==0.5.3
jdcal==1.4.1
joblib==0.16.0
Keras==2.4.3
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.18.5
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
openpyxl==3.0.5
Pillow==7.2.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pyttsx3==2.90
PyTweening==1.0.3
pywin32==228
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
xlrd==1.2.0